requirements.txt
ollama~=0.4.6
sqlalchemy
retrying~=1.3.4
python-telegram-bot~=21.10
requests
Pillow~=10.4.0
pyyaml~=6.0.2
deep_translator
easyocr~=1.7.2
opencv-python~=4.11.0.86
imageio-ffmpeg
av
ffmpeg-python
numpy~=2.1.3
redis~=5.2.1
moviepy~=2.1.2
openai-whisper~=20240930
transformers~=4.48.1
pydub~=0.25.1
whisper~=1.1.10
deep-translator~=1.11.4
torch~=2.5.1
torchaudio
speechbrain